fix : 메뉴바 수정
